requirements.txt
asgiref==3.5.2
attrs==22.1.0
brom==1.0.1b9
cached-property==1.5.2
certifi==2022.9.24
charset-normalizer==2.1.1
coreapi==2.3.3
coreschema==0.0.4
Django==4.1.2
django-cors-headers==3.13.0
django-rest-swagger==2.2.0
djangorestframework==3.14.0
drf-yasg==1.21.4
gunicorn==20.1.0
idna==3.4
inflection==0.5.1
isodate==0.6.1
itypes==1.2.0
Jinja2==3.1.2
lxml==4.9.1
MarkupSafe==2.1.1
openapi-codec==1.3.2
packaging==21.3
platformdirs==2.5.2
pyparsing==3.0.9
pytz==2022.4
requests==2.28.1
requests-file==1.5.1
requests-toolbelt==0.10.0
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
simplejson==3.17.6
six==1.16.0
sqlparse==0.4.3
uritemplate==4.1.1
urllib3==1.26.12
whitenoise==6.2.0
zeep==4.1.0
